requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
backports-abc==0.5
beautifulsoup4==4.6.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
channels==1.1.8
daphne==1.4.2
Django==1.11.4
django-allauth==0.33.0
django-autocomplete-light==3.2.10
django-livereload-server==0.2.3
django-oauth2-provider==0.2.6.1
django-picklefield==1.0.0
django-slack==5.11.1
django-webpack-loader==0.5.0
djangorestframework==3.7.7
Flask==0.12.2
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
marshmallow==2.13.5
msgpack-python==0.5.1
numpy==1.13.1
oauthlib==2.0.3
olefile==0.44
pandas==0.20.3
pigar==0.7.0
Pillow==4.2.1
pyDAL==17.8
pyee==5.0.0
python-dateutil==2.6.1
python-openid==2.2.5
pytz==2017.2
redis==2.10.6
requests==2.18.3
tornado==4.5.2
websocket-client==0.46.0
